refactor(eth-rpc): combine rpc/backend and rpc/rpcapi since they essentially one package
