--- eth/rpc/backend/client_test.go
@@ -1,1 +1,0 @@
-package backend_test
