+++ eth/rpc/rpcapi/call_tx_test.go
@@ -1,4 +1,4 @@
-package backend_test
+package rpcapi_test
 
 import (
 	"math/big"
