+++ eth/rpc/rpcapi/client_test.go
@@ -1,0 +1,1 @@
+package rpcapi_test
