--- eth/rpc/rpcapi/debugapi/trace_fallback.go
@@ -1,36 +1,0 @@
-// Copyright 2016 The go-ethereum Authors
-// This file is part of the go-ethereum library.
-//
-// The go-ethereum library is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Lesser General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// The go-ethereum library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Lesser General Public License for more details.
-//
-// You should have received a copy of the GNU Lesser General Public License
-// along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
-
-//go:build !go1.5
-// +build !go1.5
-
-// no-op implementation of tracing methods for Go < 1.5.
-
-package debug
-
-import (
-	"errors"
-)
-
-func (*API) StartGoTrace(string file) error {
-	a.logger.Debug("debug_stopGoTrace", "file", file)
-	return errors.New("tracing is not supported on Go < 1.5")
-}
-
-func (*API) StopGoTrace() error {
-	a.logger.Debug("debug_stopGoTrace")
-	return errors.New("tracing is not supported on Go < 1.5")
-}
