+++ eth/rpc/rpcapi/nonce_test.go
@@ -1,4 +1,4 @@
-package backend_test
+package rpcapi_test
 
 import (
 	sdkmath "cosmossdk.io/math"
