+++ eth/rpc/rpcapi/tx_info.go
@@ -1,5 +1,5 @@
 // Copyright (c) 2023-2024 Nibi, Inc.
-package backend
+package rpcapi
 
 import (
 	"encoding/json"
